Extract shared CLI logic from sigridci.py and agents.py into cli_common.py
sigridci.py and agents.py had near-identical argument definitions and
post-parse validation/run logic. Moved the shared pieces (capability
parsing, common argparse flags, system/subsystem name validation, and
the run-and-exit flow) into cli_common.py so each entrypoint only
defines the flags and options that are actually specific to it.
